+++ mullvad-jni/src/daemon_interface.rs
@@ -1,11 +1,16 @@
 use futures::{sync::oneshot, Future};
 use mullvad_daemon::{DaemonCommandSender, ManagementCommand};
 use mullvad_types::{
-    account::AccountData, location::GeoIpLocation, relay_constraints::RelaySettingsUpdate,
-    relay_list::RelayList, settings::Settings, states::TargetState,
+    account::AccountData,
+    location::GeoIpLocation,
+    relay_constraints::RelaySettingsUpdate,
+    relay_list::RelayList,
+    settings::Settings,
+    states::{TargetState, TunnelState},
+    wireguard::KeygenEvent,
 };
 use parking_lot::Mutex;
-use talpid_types::{net::wireguard, tunnel::TunnelStateTransition};
+use talpid_types::net::wireguard;
 
 #[derive(Debug, err_derive::Error)]
 pub enum Error {
@@ -64,14 +69,12 @@ impl DaemonInterface {
         Ok(())
     }
 
-    pub fn generate_wireguard_key(&self) -> Result<()> {
+    pub fn generate_wireguard_key(&self) -> Result<KeygenEvent> {
         let (tx, rx) = oneshot::channel();
 
         self.send_command(ManagementCommand::GenerateWireguardKey(tx))?;
 
-        rx.wait()
-            .map_err(|_| Error::NoResponse)?
-            .map_err(Error::RpcError)
+        rx.wait().map_err(|_| Error::NoResponse)
     }
 
     pub fn get_account_data(&self, account_token: String) -> Result<AccountData> {
@@ -109,7 +112,7 @@ impl DaemonInterface {
         Ok(rx.wait().map_err(|_| Error::NoResponse)?)
     }
 
-    pub fn get_state(&self) -> Result<TunnelStateTransition> {
+    pub fn get_state(&self) -> Result<TunnelState> {
         let (tx, rx) = oneshot::channel();
 
         self.send_command(ManagementCommand::GetState(tx))?;