+++ mullvad-jni/src/jni_event_listener.rs
@@ -5,9 +5,11 @@ use jni::{
     AttachGuard, JNIEnv,
 };
 use mullvad_daemon::EventListener;
-use mullvad_types::{relay_list::RelayList, settings::Settings};
+use mullvad_types::{
+    relay_list::RelayList, settings::Settings, states::TunnelState, wireguard::KeygenEvent,
+};
 use std::{sync::mpsc, thread};
-use talpid_types::{tunnel::TunnelStateTransition, ErrorExt};
+use talpid_types::ErrorExt;
 
 #[derive(Debug, err_derive::Error)]
 pub enum Error {
@@ -23,7 +25,8 @@ pub enum Error {
 
 enum Event {
     RelayList(RelayList),
-    Tunnel(TunnelStateTransition),
+    Settings(Settings),
+    Tunnel(TunnelState),
 }
 
 #[derive(Clone, Debug)]
@@ -36,21 +39,27 @@ impl JniEventListener {
 }
 
 impl EventListener for JniEventListener {
-    fn notify_new_state(&self, state: TunnelStateTransition) {
+    fn notify_new_state(&self, state: TunnelState) {
         let _ = self.0.send(Event::Tunnel(state));
     }
 
-    fn notify_settings(&self, _: Settings) {}
+    fn notify_settings(&self, settings: Settings) {
+        let _ = self.0.send(Event::Settings(settings));
+    }
 
     fn notify_relay_list(&self, relay_list: RelayList) {
         let _ = self.0.send(Event::RelayList(relay_list));
     }
+
+    // TODO: manage key events properly
+    fn notify_key_event(&self, _key_event: KeygenEvent) {}
 }
 
 struct JniEventHandler<'env> {
     env: AttachGuard<'env>,
     mullvad_ipc_client: JObject<'env>,
     notify_relay_list_event: JMethodID<'env>,
+    notify_settings_event: JMethodID<'env>,
     notify_tunnel_event: JMethodID<'env>,
     events: mpsc::Receiver<Event>,
 }
@@ -98,17 +107,24 @@ impl<'env> JniEventHandler<'env> {
             "notifyRelayListEvent",
             "(Lnet/mullvad/mullvadvpn/model/RelayList;)V",
         )?;
+        let notify_settings_event = Self::get_method_id(
+            &env,
+            &class,
+            "notifySettingsEvent",
+            "(Lnet/mullvad/mullvadvpn/model/Settings;)V",
+        )?;
         let notify_tunnel_event = Self::get_method_id(
             &env,
             &class,
             "notifyTunnelStateEvent",
-            "(Lnet/mullvad/mullvadvpn/model/TunnelStateTransition;)V",
+            "(Lnet/mullvad/mullvadvpn/model/TunnelState;)V",
         )?;
 
         Ok(JniEventHandler {
             env,
             mullvad_ipc_client,
             notify_relay_list_event,
+            notify_settings_event,
             notify_tunnel_event,
             events,
         })
@@ -128,17 +144,20 @@ impl<'env> JniEventHandler<'env> {
         while let Ok(event) = self.events.recv() {
             match event {
                 Event::RelayList(relay_list) => self.handle_relay_list_event(relay_list),
+                Event::Settings(settings) => self.handle_settings(settings),
                 Event::Tunnel(tunnel_event) => self.handle_tunnel_event(tunnel_event),
             }
         }
     }
 
     fn handle_relay_list_event(&self, relay_list: RelayList) {
+        let java_relay_list = self.env.auto_local(relay_list.into_java(&self.env));
+
         let result = self.env.call_method_unchecked(
             self.mullvad_ipc_client,
             self.notify_relay_list_event,
             JavaType::Primitive(Primitive::Void),
-            &[JValue::Object(relay_list.into_java(&self.env))],
+            &[JValue::Object(java_relay_list.as_obj())],
         );
 
         if let Err(error) = result {
@@ -149,12 +168,32 @@ impl<'env> JniEventHandler<'env> {
         }
     }
 
-    fn handle_tunnel_event(&self, event: TunnelStateTransition) {
+    fn handle_settings(&self, settings: Settings) {
+        let java_settings = self.env.auto_local(settings.into_java(&self.env));
+
+        let result = self.env.call_method_unchecked(
+            self.mullvad_ipc_client,
+            self.notify_settings_event,
+            JavaType::Primitive(Primitive::Void),
+            &[JValue::Object(java_settings.as_obj())],
+        );
+
+        if let Err(error) = result {
+            log::error!(
+                "{}",
+                error.display_chain_with_msg("Failed to call MullvadDaemon.notifySettingsEvent")
+            );
+        }
+    }
+
+    fn handle_tunnel_event(&self, event: TunnelState) {
+        let java_tunnel_state = self.env.auto_local(event.into_java(&self.env));
+
         let result = self.env.call_method_unchecked(
             self.mullvad_ipc_client,
             self.notify_tunnel_event,
             JavaType::Primitive(Primitive::Void),
-            &[JValue::Object(event.into_java(&self.env))],
+            &[JValue::Object(java_tunnel_state.as_obj())],
         );
 
         if let Err(error) = result {