+++ mullvad-types/src/lib.rs
@@ -27,11 +27,14 @@ pub use crate::custom_tunnel::*;
 #[serde(rename_all = "snake_case")]
 pub enum DaemonEvent {
     /// The daemon transitioned into a new state.
-    StateTransition(talpid_types::tunnel::TunnelStateTransition),
+    TunnelState(states::TunnelState),
 
     /// The daemon settings changed.
     Settings(settings::Settings),
 
     /// The daemon got an updated relay list.
     RelayList(relay_list::RelayList),
+
+    /// Key event
+    WireguardKey(wireguard::KeygenEvent),
 }