+++ CHANGELOG.md
@@ -23,7 +23,11 @@ Line wrap the file at 100 chars.                                              Th
 
 
 ## [Unreleased]
+### Added
+- Add a switch to turn off system notifications under Preferences in the GUI.
 
+#### Windows
+- Add migration logic to restore lost settings after major Windows update.
 
 ## [2019.5] - 2019-06-17
 ### Added